requirements.txt
Dockerfile
discord.py==2.3.2
redis==5.0.1
python-dotenv==1.0.0